.claude/skills/e2e-test-generator/requirements.txt
# @spec T002-e2e-test-generator
# Python dependencies for e2e-test-generator skill

# Core dependencies
PyYAML>=6.0.2           # YAML parsing
Jinja2>=3.1.4           # Template engine
jsonschema>=4.23.0      # JSON schema validation

# Testing dependencies
pytest>=8.3.4           # Testing framework
pytest-cov>=6.0.0       # Coverage plugin
pytest-mock>=3.14.0     # Mocking plugin

# Code quality
black>=24.10.0          # Code formatter
pylint>=3.3.2           # Linter
mypy>=1.13.0            # Type checker

# Type stubs
types-PyYAML>=6.0.12    # PyYAML type stubs
